Keep eval overrides out of prod imports
