+++ redis_sre_agent/tools/mcp/provider.py
@@ -17,7 +17,7 @@ from mcp.client.stdio import stdio_client
 from mcp.client.streamable_http import streamablehttp_client
 
 from redis_sre_agent.core.config import MCPServerConfig, MCPToolConfig
-from redis_sre_agent.evaluation.injection import get_active_mcp_runtime
+from redis_sre_agent.core.runtime_overrides import get_active_mcp_runtime
 from redis_sre_agent.tools.models import Tool, ToolCapability, ToolDefinition, ToolMetadata
 from redis_sre_agent.tools.protocols import ToolProvider
